wip: clean up moe examples
